+++ src/test/mock/SuiteOutputChannel.ts
@@ -60,4 +60,8 @@ export default class SuiteOutputChannel implements OutputChannel {
       return this.content[len - 1];
     }
   }
+
+  getLineCount(): number {
+    return this.content.length;
+  }
 }