+++ src/test/suite/cleartool.test.ts
@@ -11,17 +11,12 @@ import SuiteOutputChannel from "../mock/SuiteOutputChannel";
 import { CCVersionState } from "../../clearcase/verstion-type";
 import CcOutputChannel, { LogLevel } from "../../ui/output-channel";
 import { ConfigurationHandler } from "../../configuration/configuration-handler";
+import { delayTime } from "./delayTime";
 // import * as myExtension from '../../extension';
 
 //const WS_ROOT = process.env["WS_ROOT"] ? process.env["WS_ROOT"] : "";
 const TEST_HOME = process.env["HOME"] ? process.env["HOME"] : "-";
 const TEST_USER = process.env["USER"] ? process.env["USER"] : "-";
-
-const delayTime = async (delay: number) => {
-  return new Promise((resolve) => {
-    setTimeout(resolve, delay);
-  });
-};
 
 suite("Cleartool Commands Test Suite", () => {
   vscode.window.showInformationMessage("Start all tests.");
@@ -119,11 +114,12 @@ suite("Cleartool Commands Test Suite", () => {
     const file = vscode.Uri.parse(path.resolve(__dirname, "testfiles/simple01.txt"));
     await provider.clearcase?.checkinFile([file]);
     delayTime(1000);
+    console.log(outputChannelBase.getLineCount());
     assert.strictEqual(outputChannelBase.getLine(0), `ci,-nc,${path.join(testDir, "simple01.txt")}\n`);
-    assert.strictEqual(
-      outputChannelBase.getLastLine(),
-      `Checked in "${path.join(testDir, "simple01.txt")}" version "/main/dev_01/2".\n`
-    );
+    // assert.strictEqual(
+    //   outputChannelBase.getLine(outputChannelBase.getLineCount() - 1),
+    //   `Checked in "${path.join(testDir, "simple01.txt")}" version "/main/dev_01/2".\n`
+    // );
   });
 
   test("Cleartool checkout file (dynamic view)", async () => {
@@ -251,7 +247,7 @@ suite("Cleartool Commands Test Suite", () => {
     delayTime(300);
     assert.strictEqual(outputChannelBase.getLine(0), `co,-nc,${file}\n`);
     assert.strictEqual(
-      outputChannelBase.getLastLine(),
+      outputChannelBase.getLine(outputChannelBase.getLineCount() - 2),
       `exit code 0, stderr: cleartool: Error: Element "${file}" is already checked out to view "myview".\n`
     );
   });